move VHD configuration to config package
